+++ libfdisk/src/bsd.c
@@ -838,8 +838,8 @@ int fdisk_bsd_link_partition(struct fdisk_context *cxt)
 	fdisk_label_set_changed(cxt->label, 1);
 
 	fdisk_sinfo(cxt, FDISK_INFO_SUCCESS,
-			_("BSD partition '%c' linked to DOS partition %d."),
-			'a' + (int) i, (int) k + 1);
+			_("BSD partition '%c' linked to DOS partition %zu."),
+			'a' + (int) i, k + 1);
 	return 0;
 }
